finsih section 9 (authorization)

+++ backend/routes/posts.js
@@ -41,7 +41,8 @@ router.post(
   const post = new Post({
     title: req.body.title,
     content: req.body.content,
-    imagePath: `${url}/images/${req.file.filename}`
+    imagePath: `${url}/images/${req.file.filename}`,
+    creator: req.userData.userId
   });
 
   post.save().then(createdPost => {
@@ -103,19 +104,27 @@ router.put(
     _id: req.body.id,
     title: req.body.title,
     content: req.body.content,
-    imagePath: imagePath
+    imagePath: imagePath,
+    creator: req.userData.userId
   });
 
-  Post.updateOne({ _id: req.params.id }, post).then(result => {
-    res.status(200).json({ message: 'Update successful' });
+  Post.updateOne({ _id: req.params.id, creator: req.userData.userId }, post).then(result => {
+    if (result.nModified > 0) {
+      res.status(200).json({ message: 'Update successful' });
+    } else {
+      res.status(401).json({ message: 'Not Authorized' });
+    }
   });
 });
 
 router.delete('/:id', checkAuth, (req, res, next) => {
-  console.log(req.params.id);
-  Post.deleteOne({ _id: req.params.id }).then(result => {
-    console.log(result);
-    res.status(200).json({ message: 'Post Deleted!' });
+
+  Post.deleteOne({ _id: req.params.id, creator: req.userData.userId }).then(result => {
+    if (result.n > 0) {
+      res.status(200).json({ message: 'Delete successful' });
+    } else {
+      res.status(401).json({ message: 'Not Authorized' });
+    }
   });
 });
